MAINT move deprecated into deprecation.py
